+++ src/main/java/net/mohron/skyclaims/command/argument/Arguments.java
@@ -46,10 +46,6 @@ public final class Arguments {
     return new SchematicArgument(key);
   }
 
-  public static SortArgument sort(Text key) {
-    return new SortArgument(key);
-  }
-
   public static TargetArgument target(Text key) {
     return new TargetArgument(key);
   }